AI3604 计算机视觉/Project/Experiment/requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
einops==0.8.0
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
huggingface-hub==0.23.1
humanize==4.9.0
icecream==2.1.3
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
monai==1.3.1
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nptyping==2.5.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
Pygments==2.18.0
pynrrd==1.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
requests==2.32.1
rich==13.7.1
safetensors==0.4.3
scikit-image==0.18.3
scipy==1.13.0
seaborn==0.13.2
segment-anything==1.0
shellingham==1.5.4
SimpleITK==2.3.1
six==1.16.0
slicerio==1.0.0
sympy==1.12
tifffile==2024.5.10
timm==1.0.3
torch==2.3.0
torchio==0.19.6
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
